feat: :sparkles: Add lexer and validate tokens
tokens에 있는 token들이 유효한 순서로 나왔는지 미리 검토

closes #43
